Disable the ocaml detection for the dune port
This is used to determine whether the ocamlbindings should be built.
Given we're specifically in the dune port there's no point in running
this as we obviously want the ocaml bindings.

It's probably worth working on a better version by making this
configurable so that the dune rules always enables them or to eventually
make the detection more robust in the future.

Signed-off-by: Nathan Rebours <nathan.p.rebours@gmail.com>